+++ QRealFourier-master/code/sources/qfouriervariablecalculator.cpp
@@ -60,12 +60,12 @@ void QFourierVariableCalculator::setSize(int size)
 	QFourierCalculator::setSize(size);
 	if(mFourierTransform == 0)
 	{
-		mFourierTransform = new ffft::FFTReal<float>(mSize);
+		mFourierTransform = new ffft::FFTReal<double>(mSize);
 	}
 	else if(mFourierTransform->get_length() != mSize)
 	{
 		delete mFourierTransform;
-		mFourierTransform = new ffft::FFTReal<float>(mSize);
+		mFourierTransform = new ffft::FFTReal<double>(mSize);
 	}
 }
